+++ CHANGELOG.md
@@ -4,7 +4,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- Nothing yet
+### Fixed
+- sync tasks support
+### Added
+- aliases for event (un)registering
+- "once" event
 
 ## [1.0.1] 2018-10-23
 ### Added